allow any size for values
This is important in order to allow Ancestral values (in particular
function parameters) to have any size (which in turn the code will
partition/select into 1/2/4/8 bytes chunks).

+++ src/mil.rs
@@ -50,11 +50,11 @@ pub type Index = u16;
 #[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
 pub enum RegType {
     Effect,
-    Bytes(u8),
+    Bytes(usize),
     Bool,
 }
 impl RegType {
-    pub(crate) fn bytes_size(&self) -> Option<u8> {
+    pub(crate) fn bytes_size(&self) -> Option<usize> {
         match self {
             RegType::Effect => None,
             RegType::Bool => None,
@@ -164,7 +164,7 @@ pub enum Insn {
     /// Exists purely to give the phi node an index that the rest of the program can refer to (in
     /// SSA).
     Phi {
-        size: u8,
+        size: usize,
     },
     PhiBool,
